Create Binary-Search.c
Description: In the program above, we declare I, low, high, mid, n, key, array [size].
•	i is used for iteration through the array.
•	low is used to hold the first array index.
•	high is used to hold the last array index.
•	mid holds the middle value calculated by adding high and low and dividing it by 2.
•	n is the number of elements in the array.
•	key is the element to search.
•	array is the array element of size 100.
Discussion: We first, take in the number of elements the user array needs and store it in n. Next, we take the elements from the user. A for loop is used for this process. Then, we take the number to be searched from the array and store it in the key. Next, we assign 0 to the low variable which is the first index of an array and n-1 to the high element, which is the last element in the array. We then calculate the mid value.    mid = (low+high)/2 to get the middle index of the array.